docs: publish data plane public api into dedicated swagger hub page
