--- extensions/data-plane/data-plane-api/src/main/resources/META-INF/services/org.eclipse.edc.spi.system.ServiceExtension
@@ -1,1 +1,0 @@
-org.eclipse.edc.connector.dataplane.api.DataPlaneApiExtension
