+++ extensions/data-plane/data-plane-control-api/src/main/java/org/eclipse/edc/connector/dataplane/api/DataPlaneControlApiExtension.java
@@ -1,0 +1,53 @@
+/*
+ *  Copyright (c) 2022 Microsoft Corporation
+ *
+ *  This program and the accompanying materials are made available under the
+ *  terms of the Apache License, Version 2.0 which is available at
+ *  https://www.apache.org/licenses/LICENSE-2.0
+ *
+ *  SPDX-License-Identifier: Apache-2.0
+ *
+ *  Contributors:
+ *       Microsoft Corporation - initial API and implementation
+ *       Mercedes-Benz Tech Innovation GmbH - publish public api context into dedicated swagger hub page
+ *
+ */
+
+package org.eclipse.edc.connector.dataplane.api;
+
+import org.eclipse.edc.connector.api.control.configuration.ControlApiConfiguration;
+import org.eclipse.edc.connector.dataplane.api.controller.DataPlaneControlApiController;
+import org.eclipse.edc.connector.dataplane.spi.manager.DataPlaneManager;
+import org.eclipse.edc.runtime.metamodel.annotation.Extension;
+import org.eclipse.edc.runtime.metamodel.annotation.Inject;
+import org.eclipse.edc.spi.system.ServiceExtension;
+import org.eclipse.edc.spi.system.ServiceExtensionContext;
+import org.eclipse.edc.web.spi.WebService;
+
+/**
+ * This extension provides set of endpoints to trigger/monitor/cancel data transfers that should be accessible only
+ * by the Control Plane.
+ */
+@Extension(value = DataPlaneControlApiExtension.NAME)
+public class DataPlaneControlApiExtension implements ServiceExtension {
+    public static final String NAME = "Data Plane Control API";
+
+    @Inject
+    private DataPlaneManager dataPlaneManager;
+
+    @Inject
+    private WebService webService;
+
+    @Inject
+    private ControlApiConfiguration controlApiConfiguration;
+
+    @Override
+    public String name() {
+        return NAME;
+    }
+
+    @Override
+    public void initialize(ServiceExtensionContext context) {
+        webService.registerResource(controlApiConfiguration.getContextAlias(), new DataPlaneControlApiController(dataPlaneManager));
+    }
+}
