+++ extensions/data-plane/data-plane-control-api/src/main/resources/META-INF/services/org.eclipse.edc.spi.system.ServiceExtension
@@ -1,0 +1,1 @@
+org.eclipse.edc.connector.dataplane.api.DataPlaneControlApiExtension
