+++ launchers/data-plane-server/build.gradle.kts
@@ -9,6 +9,7 @@
  *
  *  Contributors:
  *       Microsoft Corporation - initial API and implementation
+ *       Mercedes-Benz Tech Innovation GmbH - publish public api context into dedicated swagger hub page
  *
  */
 
@@ -26,7 +27,7 @@ dependencies {
     implementation(project(":extensions:common:http"))
     implementation(project(":extensions:common:configuration:configuration-filesystem"))
     implementation(project(":extensions:data-plane:data-plane-http"))
-    implementation(project(":extensions:data-plane:data-plane-api"))
+    implementation(project(":extensions:data-plane:data-plane-control-api"))
 }
 
 application {
